requirements/base.txt
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
uvicorn==0.21.1